refactor(frontend): consolidate user-summary mapping
Six api-client modules rebuilt the same {id, login, displayName, deleted,
isBot, avatarUrl} block from the generated User protobuf by hand, and
memberDirectory and notifications duplicated the customStatus timestamp
conversion verbatim. Extract a leaf userSummary module with mapUserSummary,
mapOptionalUserSummary, and mapUserPresenceView (type-only protobuf imports,
so runtime chunk graphs stay put).

- hooks.ts: UserSummaryForCache is now an alias of the shared UserSummary.
- notifications.ts: notificationActor maps via the shared helpers.
- memberDirectory.ts: DirectoryMember extends the shared types; blank-member
  default row and isBot normalization to false are preserved explicitly.
- roomTimeline.ts: cache priming uses the shared mapper, including priming
  straight from batchGetUsers summaries instead of rebuilding them.

Deleted-user rendering stays surface-local (timeline keeps its Deleted User
fallback); adminUsers keeps its own shapes where field subsets genuinely
differ. No behavior change; existing specs catch the directory's isBot
normalization.
