+++ apps/frontend/src/lib/api-client/users.ts
@@ -1,21 +1,14 @@
 import { authHeaders, createChattoClient } from './connect.js';
 import { UserService } from '@chatto/api-types/api/v1/member_directory_connect';
 import type { DirectoryMember as APIDirectoryMember } from '@chatto/api-types/api/v1/member_directory_pb';
-import type { User as APIUser } from '@chatto/api-types/api/v1/users_pb';
+
+export { mapUserSummary, mapOptionalUserSummary, type UserSummary } from './userSummary.js';
+import { mapUserSummary, type UserSummary } from './userSummary.js';
 
 export type UserAPIConfig = {
   baseUrl: string;
   bearerToken: string | null;
   onAuthenticationRequired?: (serverId: string) => void;
-};
-
-export type UserSummary = {
-  id: string;
-  login: string;
-  displayName: string;
-  deleted: boolean;
-  isBot?: boolean;
-  avatarUrl: string | null;
 };
 
 export function createUserAPI(config: UserAPIConfig) {
@@ -37,15 +30,4 @@ export type UserAPI = ReturnType<typeof createUserAPI>;
 
 export function mapDirectoryMemberUserSummary(member: APIDirectoryMember): UserSummary | null {
   return member.user ? mapUserSummary(member.user) : null;
-}
-
-export function mapUserSummary(user: APIUser): UserSummary {
-  return {
-    id: user.id,
-    login: user.login,
-    displayName: user.displayName,
-    deleted: user.deleted,
-    isBot: user.isBot,
-    avatarUrl: user.avatarUrl || null
-  };
 }
